Split out a common ColumnStoreSpec, and create InMemoryColumnStoreSpec
