+++ core/src/test/scala/filodb.core/store/InMemoryColumnStoreSpec.scala
@@ -1,0 +1,6 @@
+package filodb.core.store
+
+class InMemoryColumnStoreSpec extends ColumnStoreSpec {
+  import scala.concurrent.ExecutionContext.Implicits.global
+  val colStore = new InMemoryColumnStore
+}
